Fix the schedule header column on devices with low height
On devices with a low height, the fixed positioned header column might take up more than half the screen size, making the schedule unusable on those devices.

This fix will also set that large padding - why is that there in the first place - if the `min-height` is at least `1000px`, fixing that issue on those devices.